% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_args}
\alias{add_args}
\title{Add platform-specific additional arguments to a GitHub workflow YAML file}
\usage{
add_args(workflow_name, additional_args, txt = NULL, prev_line = NULL)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{additional_args}{A named list of additional command line arguments to be
passed to the workflow. The names of the list represent the platforms (windows,
macos, or ubuntu), and the values are character vectors of arguments.
These additional arguments are executed after the step that checks out the
repository and before the step that sets up R.}

\item{txt}{A character vector of YAML file lines. If NULL, the function reads the YAML file
automatically. Default is NULL.}

\item{prev_line}{An integer specifying the line number in txt where the arguments
should be added. If NULL, the function determines this value automatically. Default is NULL.}
}
\description{
Add platform-specific additional arguments to a GitHub workflow YAML file
}
