% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_public_rspm_false}
\alias{add_public_rspm_false}
\title{Add public rspm is false option to workflows.}
\usage{
add_public_rspm_false(uses_line, gha)
}
\arguments{
\item{uses_line}{text that includes "uses: "}

\item{gha}{The workflow file that has been read in using readLines}
}
\value{
The modified workflow file
}
\description{
Add public rspm is false option to workflows.
}
