% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_quarto_true}
\alias{add_quarto_true}
\title{Add quarto is true option to workflows.}
\usage{
add_quarto_true(uses_line, gha)
}
\arguments{
\item{uses_line}{text that includes "uses: "}

\item{gha}{The workflow file that has been read in using readLines}
}
\value{
The modified workflow file
}
\description{
Add quarto is true option to workflows.
}
