% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_workflow_name}
\alias{check_workflow_name}
\title{Check the workflow name is formatted correctly}
\usage{
check_workflow_name(workflow_name)
}
\arguments{
\item{workflow_name}{What to name the github actionworkflow locally. Must
have the file extension \code{.yml}}
}
\description{
Basically check that it is a filename that ends in .yml
}
