% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_caller_template}
\alias{copy_caller_template}
\title{Copy the caller template to an R package}
\usage{
copy_caller_template(
  template_name = "call-spell-check.yml",
  workflow_name = "call-spell-check.yml"
)
}
\arguments{
\item{template_name}{The name of the template in the ghactions4r package}

\item{workflow_name}{The user defined name of the workflow to create}
}
\value{
The path to the caller workflow file copied
}
\description{
Copy the caller template to an R package
}
