% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tags}
\alias{get_tags}
\title{Get the tags by calling the GitHub rest API}
\usage{
get_tags(repo = "nmfs-ost/ghactions4r", headers = NULL)
}
\arguments{
\item{repo}{owner/repo}

\item{headers}{NULL for now}
}
\description{
Get the tags by calling the GitHub rest API
}
