% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_build_deploy_bookdown}
\alias{use_build_deploy_bookdown}
\title{use workflow in current pkg to build and deploy (update) bookdown
Builds the bookdown, then deploys it to a branch in the same repository called gh-pages.
The repository must be}
\usage{
use_build_deploy_bookdown(
  workflow_name = "call-build-deploy-bookdown.yml",
  bookdown_input = ".",
  bookdown_output_dir = "_book",
  deployment_dir = "_book"
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{bookdown_input}{Used as the input for \code{input} in \code{bookdown::render_book()}.}

\item{bookdown_output_dir}{Used as the input for \code{output_dir} in \code{bookdown::render_book()}.
Note that \code{NULL} will currently not work.}

\item{deployment_dir}{The directory to deploy to on the gh-pages branch.}
}
\value{
Path to the .yml file.
}
\description{
use workflow in current pkg to build and deploy (update) bookdown
Builds the bookdown, then deploys it to a branch in the same repository called gh-pages.
The repository must be
}
