% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_calc_cov_summaries}
\alias{use_calc_cov_summaries}
\title{Workflow for calculating coverage and summarizing using octocov toolkit}
\usage{
use_calc_cov_summaries(
  workflow_name = "call-calc-cov-summaries.yml",
  build_trigger = c("manually", "pull_request", "push_to_main", "push_to_all_branches"),
  use_public_rspm = TRUE,
  depends_on_quarto = FALSE
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{build_trigger}{Select the build trigger. Options may include (depending on the workflow)
running on pushing commits to main ("push_to_main", the default); run when a pull request is
opened, reopened, or updated ("pull_request"); run when a comment containing
the command \verb{\\doc-and-style} is made on a pull request by people with write permissions
on the repository ("pr_comment"); run manually with the
workflow_dispatch trigger ("manually"); run on the default branch (usually
main) once a week ("weekly"). Currently, it is not possible to create a file
with multiple build triggers from the function. Instead, users should add all of the
appropriate build triggers by editing the file after the use_* function.}

\item{use_public_rspm}{Use posit package manager instead of CRAN to install dependencies? The
advantage here is that dependencies are precompiled, so install should be much quicker. In
rare situations (like packages with TMB dependencies), using use_public_rspm = FALSE may be
a better option. Note a setting only needs to be specified in the yml if use_public_rspm is FALSE, so
there will be no setting added if use_public_rspm is TRUE.}

\item{depends_on_quarto}{Adds an option that installs quarto as a dependency.}
}
\description{
This workflow calculates coverage using the covr package, then creates
summaries to post as a GitHub action summary or as pull request comment. All
data remains in the GitHub repository.
}
