% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_connect_publish}
\alias{use_connect_publish}
\title{use workflow to publish to Posit Connect from a private repo}
\usage{
use_connect_publish(workflow_name = "call-connect-publish.yml")
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}
}
\value{
The path to the new github action file.
}
\description{
This action publishes to connect using the git-backed workflow. Note that
it is only necessary to do it this way if the product is in a private or
internal repository; otherwise, it is far easier to deploy following the
instructions in https://docs.posit.co/connect/user/git-backed/
}
\details{
To use this workflow, the user will need to generate the manifest.json, and add 2 secrets, one named
CONNECT_URL containing the connect url (complete with https:// in front of
the address) and one named CONNECT_API_KEY, containing an API Key from connect.
To generate the manifest file, follow instructions in the create a manifest
from r setup and pushing it to github:
https://docs.posit.co/connect/user/git-backed/#creating-a-manifest-file-from-r
See  https://docs.posit.co/connect/user/api-keys/#api-keys-creating for
instructions on creating and API key from connect and
https://octopus.com/blog/githubactions-secrets for instructions on creating
secrets.
}
