% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_style_description}
\alias{use_style_description}
\title{use workflow in current pkg to run \code{usethis::use_tidy_description()}}
\usage{
use_style_description(workflow_name = "call-style-description.yml")
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}
}
\description{
Run \code{usethis::use_tidy_description()} upon changes to the DESCRIPTION file
and submit results as a PR.
}
