% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_style_r_code}
\alias{use_style_r_code}
\title{use workflow in current pkg to run styler::style_pkg() and submit results as a PR}
\usage{
use_style_r_code(workflow_name = "call-style.yml")
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}
}
\description{
use workflow in current pkg to run styler::style_pkg() and submit results as a PR
}
