% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_update_pkgdown}
\alias{use_update_pkgdown}
\title{Creates a workflow in current R package to update an existing pkgdown GitHub pages site}
\usage{
use_update_pkgdown(
  workflow_name = "call-update-pkgdown.yml",
  build_trigger = "push_to_main",
  additional_args = NULL,
  tag_ghactions4r = NULL
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{build_trigger}{Select the
\href{https://docs.github.com/en/actions/get-started/understand-github-actions#events}{event that triggers the workflow}. Options are:
\describe{
\item{\code{push_to_main}}{Run on pushing commits to main (default)}
\item{\code{weekly}}{Run on the default branch (usually main) once a week}
\item{\code{manually}}{run manually with the workflow_dispatch trigger}
}
Multiple build triggers are allowed; specify them as a vector. Note that
invalid build triggers will be silently removed as long as one build trigger
is specified correctly.}

\item{additional_args}{A named list of additional command line arguments to be
passed to the workflow. The names of the list represent the platforms (windows,
macos, or ubuntu), and the values are character vectors of arguments.
These additional arguments are executed after the step that checks out the
repository and before the step that sets up R.}

\item{tag_ghactions4r}{Tag of ghactions4r workflow to pin to for the
reusable workflow. This ensures compatibility with a specific version of
ghactions4r functions. The tradeoff is that fixes will not happen
automatically; instead, the user will have to update the pinned version of
the caller workflow used. See the options by looking at the
\href{https://github.com/nmfs-ost/ghactions4r/tags}{ghactions4r tags}. Another
option is to pin to commit hashes, but the user will need to do this by
manually editing the caller workflow after creation. See
\href{https://docs.github.com/en/actions/reference/security/secure-use}{github documentation on security}
to understand how this is more secure. Again, the tradeoff is that the
workflow will not be automatically updated.}
}
\description{
For more information, see \href{https://nmfs-fish-tools.github.io/ghactions4r/articles/set_up.html}{the setup details vignette} before
running this function. This workflow will assume the website is built from
a branch called \code{gh-pages}.
}
\examples{
\dontrun{
use_update_pkgdown(
  additional_args = list(
    ubuntu = c(
      "sudo apt-get update",
      "sudo apt-get install -y libcurl4-openssl-dev",
      "sudo add-apt-repository ppa:ubuntu-toolchain-r/test",
      "sudo apt-get install --only-upgrade libstdc++6"
    ),
    macos = c("brew install curl")
  )
)
}
}
