% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_update_roxygen_docs}
\alias{use_update_roxygen_docs}
\title{use workflow in curent pkg to run devtools::document() and submit results as a PR}
\usage{
use_update_roxygen_docs()
}
\description{
use workflow in curent pkg to run devtools::document() and submit results as a PR
}
