% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_version_pin}
\alias{use_version_pin}
\title{Pin to a particular version of ghactions4r rather than main}
\usage{
use_version_pin(tag, gha)
}
\arguments{
\item{tag}{Version of ghactions4r to pin to a specific version}

\item{gha}{The workflow text}
}
\description{
Pin to a particular version of ghactions4r rather than main
}
