% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{validate_additional_args}
\alias{validate_additional_args}
\title{Validate additional arguments for R functions}
\usage{
validate_additional_args(additional_args)
}
\arguments{
\item{additional_args}{A named list of additional command line arguments to be
passed to the workflow. The names of the list represent the platforms (windows,
macos, or ubuntu), and the values are character vectors of arguments.
These additional arguments are executed after the step that checks out the
repository and before the step that sets up R.}
}
\description{
Validate additional arguments for R functions
}
