% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_build_pkgdown}
\alias{use_build_pkgdown}
\title{use workflow in current pkg to check pkgdown site builds.}
\usage{
use_build_pkgdown(
  workflow_name = "call-build-pkgdown.yml",
  additional_args = NULL
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{additional_args}{A named list of additional command line arguments to be
passed to the workflow. The names of the list represent the platforms (windows,
macos, or ubuntu), and the values are character vectors of arguments.
These additional arguments are executed after the step that checks out the
repository and before the step that sets up R.}
}
\description{
This workflow will try to build the pkgdown site, and store it as an artifact
to be retrieved from the GitHub actions run. It will fail if the pkgdown site
cannot be built. This is a companion to use_update_pkgdown(), as the update_pkgdown
the site, and therefore can be used to test if the build is working in cases
where you do not want to deploy as well.
}
\examples{
\dontrun{
use_build_pkgdown(
  additional_args = list(
    ubuntu = c(
      "sudo apt-get update",
      "sudo apt-get install -y libcurl4-openssl-dev",
      "sudo add-apt-repository ppa:ubuntu-toolchain-r/test",
      "sudo apt-get install --only-upgrade libstdc++6"
    ),
    macos = c("brew install curl")
  )
)
}
}
