% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_build_pkgdown}
\alias{use_build_pkgdown}
\title{use workflow in current pkg to check pkgdown site builds.}
\usage{
use_build_pkgdown(workflow_name = "call-build-pkgdown.yml")
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}
}
\description{
use workflow in current pkg to check pkgdown site builds.
}
