% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_build_pkgdown}
\alias{use_build_pkgdown}
\title{use workflow in current pkg to check pkgdown site builds.}
\usage{
use_build_pkgdown(
  workflow_name = "call-build-pkgdown.yml",
  build_trigger = c("manually", "weekly", "pull_request", "push_to_main",
    "push_to_all_branches"),
  additional_args = NULL
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{build_trigger}{Select the build trigger. Options may include (depending on the workflow)
running on pushing commits to main ("push_to_main", the default); run when a pull request is
opened, reopened, or updated ("pull_request"); run when a comment containing
the command \verb{\\doc-and-style} is made on a pull request by people with write permissions
on the repository ("pr_comment"); run manually with the
workflow_dispatch trigger ("manually"); run on the default branch (usually
main) once a week ("weekly"). Currently, it is not possible to create a file
with multiple build triggers from the function. Instead, users should add all of the
appropriate build triggers by editing the file after the use_* function.}

\item{additional_args}{A named list of additional command line arguments to be
passed to the workflow. The names of the list represent the platforms (windows,
macos, or ubuntu), and the values are character vectors of arguments.
These additional arguments are executed after the step that checks out the
repository and before the step that sets up R.}
}
\description{
This workflow will try to build the pkgdown site, and store it as an artifact
to be retrieved from the GitHub actions run. It will fail if the pkgdown site
cannot be built. This is a companion to use_update_pkgdown(), as the update_pkgdown
the site, and therefore can be used to test if the build is working in cases
where you do not want to deploy as well.
}
\examples{
\dontrun{
use_build_pkgdown(
  additional_args = list(
    ubuntu = c(
      "sudo apt-get update",
      "sudo apt-get install -y libcurl4-openssl-dev",
      "sudo add-apt-repository ppa:ubuntu-toolchain-r/test",
      "sudo apt-get install --only-upgrade libstdc++6"
    ),
    macos = c("brew install curl")
  )
)
}
}
