% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_calc_cov_summaries}
\alias{use_calc_cov_summaries}
\title{Workflow for calculating coverage and summarizing using octocov toolkit
This workflow calculates coverage using the covr package, then creates
summaries to post as a GitHub action summary or as pull request comment. All
data remains in the GitHub repository.}
\usage{
use_calc_cov_summaries(
  workflow_name = "call-calc-cov-summaries.yml",
  use_public_rspm = TRUE
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{use_public_rspm}{Use posit package manager instead of CRAN to install dependencies? The
advantage here is that dependencies are precompiled, so install should be much quicker. In
rare situations (like packages with TMB dependencies), using use_public_rspm = FALSE may be
a better option. Note a setting only needs to be specified in the yml if use_public_rspm is FALSE, so
there will be no setting added if use_public_rspm is TRUE.}
}
\description{
Workflow for calculating coverage and summarizing using octocov toolkit
This workflow calculates coverage using the covr package, then creates
summaries to post as a GitHub action summary or as pull request comment. All
data remains in the GitHub repository.
}
