% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_calc_coverage}
\alias{use_calc_coverage}
\title{workflow for calculating code coverage}
\usage{
use_calc_coverage(workflow_name = "call-calc-coverage.yml")
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}
}
\description{
workflow for calculating code coverage
}
