% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_create_cov_badge}
\alias{use_create_cov_badge}
\title{workflow for adding a code coverage badge}
\usage{
use_create_cov_badge(
  workflow_name = "call-create-cov-badge.yml",
  use_public_rspm = TRUE
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{use_public_rspm}{Use posit package manager instead of CRAN to install dependencies? The
advantage here is that dependencies are precompiled, so install should be much quicker. In
rare situations (like packages with TMB dependencies), using use_public_rspm = FALSE may be
a better option. Note a setting only needs to be specified in the yml if use_public_rspm is FALSE, so
there will be no setting added if use_public_rspm is TRUE.}
}
\description{
This workflow will create a new code coverage badge with the latest overall
coverage value from the main branch on every push to main. The coverage badges are pushed
to a branch called badges within the repository, which can be referenced in
the readme file on main to display the current code coverage.
}
