% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_create_cov_badge}
\alias{use_create_cov_badge}
\title{workflow for adding a code coverage badge}
\usage{
use_create_cov_badge(
  workflow_name = "call-create-cov-badge.yml",
  build_trigger = c("push_to_main", "weekly", "manually"),
  use_public_rspm = TRUE,
  depends_on_quarto = FALSE,
  tag_ghactions4r = NULL
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{build_trigger}{Select the build trigger. Options may include (depending on the workflow)
running on pushing commits to main ("push_to_main"); run when a pull request is
opened, reopened, or updated ("pull_request"); run when a comment containing
the command \verb{\\doc-and-style} is made on a pull request by people with write permissions
on the repository ("pr_comment"); run manually with the
workflow_dispatch trigger ("manually"); run on the default branch (usually
main) once a week ("weekly"). Currently, it is not possible to create a file
with multiple build triggers from the function. Instead, users should add all of the
appropriate build triggers by editing the file after the use_* function.}

\item{use_public_rspm}{Use posit package manager instead of CRAN to install dependencies? The
advantage here is that dependencies are precompiled, so install should be much quicker. In
rare situations (like packages with TMB dependencies), using use_public_rspm = FALSE may be
a better option. Note a setting only needs to be specified in the yml if use_public_rspm is FALSE, so
there will be no setting added if use_public_rspm is TRUE.}

\item{depends_on_quarto}{Adds an option that installs quarto as a dependency.}

\item{tag_ghactions4r}{Tag of ghactions4r workflow to pin to for the
reusable workflow. This ensures compatibility with a specific version of
ghactions4r functions. The tradeoff is that fixes will not happen
automatically; instead, the user will have to update the pinned version of
the caller workflow used. See the options by looking at the
\href{https://github.com/nmfs-ost/ghactions4r/tags}{ghactions4r tags}. Another
option is to pin to commit hashes, but the user will need to do this by
manually editing the caller workflow after creation. See
\href{https://docs.github.com/en/actions/reference/security/secure-use}{github documentation on security}
to understand how this is more secure. Again, the tradeoff is that the
workflow will not be automatically updated.}
}
\description{
This workflow will create a new code coverage badge with the latest overall
coverage value from the main branch on every push to main. The coverage badges are pushed
to a branch called badges within the repository, which can be referenced in
the readme file on main to display the current code coverage.
}
