% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_doc_and_style_r}
\alias{use_doc_and_style_r}
\title{Use workflow in current pkg  to automate documenting and styling}
\usage{
use_doc_and_style_r(use_rm_dollar_sign = FALSE)
}
\arguments{
\item{use_rm_dollar_sign}{in addition to devtools::document and styler::style_pkg,
should r4ss:::rm_dollar_sign be run? Defaults to FALSE.}
}
\description{
Use workflow in current pkg  to automate documenting and styling
}
