% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_doc_and_style_r}
\alias{use_doc_and_style_r}
\title{Use workflow in current pkg  to automate documenting and styling}
\usage{
use_doc_and_style_r(
  workflow_name = "call-doc-and-style-r.yml",
  use_rm_dollar_sign = FALSE,
  how_to_commit = c("pull_request", "directly"),
  build_trigger = c("push_to_main", "pull_request", "manually", "weekly"),
  use_pat = FALSE,
  pat_name = "PAT"
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{use_rm_dollar_sign}{in addition to devtools::document and
styler::style_pkg, should r4ss:::rm_dollar_sign be run? Defaults to FALSE.}

\item{how_to_commit}{Where should changes made to style and documentation be
committed? Options are 1) in a pull request to the branch ("pull_request")
where the workflow started; or 2) directly to the branch ("directly") where
the workflow started.}

\item{build_trigger}{Select the build trigger. Options are to run on pushing
commits to main ("push_to_main", the default); run when a pull request is
opened, reopened, or updated ("pull_request"); run manually with the
workflow_dispatch trigger ("manually"); run on the default branch (usually
main) once a week ("weekly").}

\item{use_pat}{Should a personal access token (PAT) stored as a github secret
be used? This is only necessary if you want the pull request generated by
the doc and style workflow to be able to start other github action runs.}

\item{pat_name}{Name of the personal access token (PAT), as stored in secrets.
Only used if \code{use_pat = TRUE}. After
\href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens}{generating the personal access token},
See \href{https://docs.github.com/en/actions/security-guides/using-secrets-in-github-actions#creating-secrets-for-a-repository}{how to add it as a secret}.}
}
\description{
Style your R package components automatically by running devtools::document(),
styler::style_pkg(), and usethis::use_tidy_description().
}
\details{
Note that allow the pull request created from this workflow can
be used to trigger github action runs, by using a personal access
token to provide the right permissions. The PAT is first
\href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens}{generated on a user account},
then \href{https://docs.github.com/en/actions/security-guides/using-secrets-in-github-actions#creating-secrets-for-a-repository}{added to the repository or organization as a secret}.
Using a PAT can cause recursive runs, so should be used
with caution. See github documentation on
\href{https://docs.github.com/en/actions/using-workflows/triggering-a-workflow#triggering-a-workflow-from-a-workflow}{triggering a workflow from a workflow}
for more information.
}
