% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_r_cmd_check}
\alias{use_r_cmd_check}
\title{Use workflow to run r cmd check on linux, mac, and windows gh actions}
\usage{
use_r_cmd_check()
}
\description{
Use workflow to run r cmd check on linux, mac, and windows gh actions
}
