% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_r_cmd_check}
\alias{use_r_cmd_check}
\title{Use workflow to run r cmd check on Linux, Mac, and Windows GitHub Actions}
\usage{
use_r_cmd_check(
  workflow_name = "call-r-cmd-check.yml",
  use_full_build_matrix = FALSE,
  depends_on_tmb = FALSE,
  depends_on_quarto = FALSE,
  additional_args = NULL
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{use_full_build_matrix}{Run R cmd check with two older versions of R in
addition to the three runs that use the release version.}

\item{depends_on_tmb}{Adds an option that install Matrix from source for windows
and Mac builds to solved a nuanced issue for packages dependent on TMB.
See this \href{https://groups.google.com/g/tmb-users/c/-GhmuuDP_OQ}{google groups thread}
for more information.}

\item{depends_on_quarto}{Adds an option that installs quarto before running r cmd
check.}

\item{additional_args}{A named list of additional command line arguments to be
passed to the workflow. The names of the list represent the platforms (windows,
macos, or ubuntu), and the values are character vectors of arguments.
These additional arguments are executed after the step that checks out the
repository and before the step that sets up R.}
}
\description{
Use workflow to run r cmd check on Linux, Mac, and Windows GitHub Actions
}
\examples{
\dontrun{
use_r_cmd_check(
  additional_args = list(
    ubuntu = c(
      "sudo apt-get update",
      "sudo apt-get install -y libcurl4-openssl-dev",
      "sudo add-apt-repository ppa:ubuntu-toolchain-r/test",
      "sudo apt-get install --only-upgrade libstdc++6"
    ),
    macos = c("brew install curl")
  )
)
}
}
