% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_spell_check}
\alias{use_spell_check}
\title{use workflow to run spelling::spell_check_package()}
\usage{
use_spell_check(workflow_name = "call-spell-check.yml")
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}
}
\value{
The path to the new github action file.
}
\description{
use workflow to run spelling::spell_check_package()
}
