% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_spell_check}
\alias{use_spell_check}
\title{use workflow to run spelling::spell_check_package()}
\usage{
use_spell_check(
  workflow_name = "call-spell-check.yml",
  build_trigger = c("manually", "weekly", "pull_request", "push_to_main",
    "push_to_all_branches"),
  spell_check_additional_files = FALSE,
  spell_check_report_level = c("error", "warning")
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{build_trigger}{Select the build trigger. Options may include (depending on the workflow)
running on pushing commits to main ("push_to_main", the default); run when a pull request is
opened, reopened, or updated ("pull_request"); run when a comment containing
the command \verb{\\doc-and-style} is made on a pull request by people with write permissions
on the repository ("pr_comment"); run manually with the
workflow_dispatch trigger ("manually"); run on the default branch (usually
main) once a week ("weekly"). Currently, it is not possible to create a file
with multiple build triggers from the function. Instead, users should add all of the
appropriate build triggers by editing the file after the use_* function.}

\item{spell_check_additional_files}{Logical. Should the workflow run \code{spell_check_files()} on specified files
in the package? Defaults to FALSE.}

\item{spell_check_report_level}{Character. The level of the report to generate for the additional spell check.
Options are "warning" or "error". Defaults to "error". If set to "error",
the workflow will fail if any spelling errors are found. If set to "warning", the
the workflow will pass even if spelling errors are found.}
}
\description{
This workflow will run the spell_check_package() function from the spelling package
on the current package, and will fail if any spelling errors are found.
It will also run the spell_check_files() on specified file types if spell_check_additional_files is TRUE.
}
