% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_update_pkgdown}
\alias{use_update_pkgdown}
\title{Creates a workflow in current R package to update an existing pkgdown GitHub pages site}
\usage{
use_update_pkgdown(
  workflow_name = "call-update-pkgdown.yml",
  additional_args = NULL
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{additional_args}{A named list of additional command line arguments to be
passed to the workflow. The names of the list represent the platforms (windows,
macos, or ubuntu), and the values are character vectors of arguments.
These additional arguments are executed after the step that checks out the
repository and before the step that sets up R.}
}
\description{
For more information, see \href{https://nmfs-fish-tools.github.io/ghactions4r/articles/set_up.html}{the setup details vignette} before
running this function. This workflow will assume the website is built from
a branch called \code{gh-pages}.
}
\examples{
\dontrun{
use_update_pkgdown(
  additional_args = list(
    ubuntu = c(
      "sudo apt-get update",
      "sudo apt-get install -y libcurl4-openssl-dev",
      "sudo add-apt-repository ppa:ubuntu-toolchain-r/test",
      "sudo apt-get install --only-upgrade libstdc++6"
    ),
    macos = c("brew install curl")
  )
)
}
}
