% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_update_pkgdown}
\alias{use_update_pkgdown}
\title{use workflow in current pkg to update pkg down, where the site is deployed to a branch called gh-pages}
\usage{
use_update_pkgdown(workflow_name = "call-update-pkgdown.yml")
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}
}
\description{
use workflow in current pkg to update pkg down, where the site is deployed to a branch called gh-pages
}
