% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_update_pkgdown}
\alias{use_update_pkgdown}
\title{Creates a workflow in current R package to update an existing pkgdown GitHub pages site}
\usage{
use_update_pkgdown(
  workflow_name = "call-update-pkgdown.yml",
  build_trigger = c("push_to_main", "manually", "weekly"),
  additional_args = NULL
)
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}

\item{build_trigger}{Select the build trigger. Options may include (depending on the workflow)
running on pushing commits to main ("push_to_main", the default); run when a pull request is
opened, reopened, or updated ("pull_request"); run when a comment containing
the command \verb{\\doc-and-style} is made on a pull request by people with write permissions
on the repository ("pr_comment"); run manually with the
workflow_dispatch trigger ("manually"); run on the default branch (usually
main) once a week ("weekly"). Currently, it is not possible to create a file
with multiple build triggers from the function. Instead, users should add all of the
appropriate build triggers by editing the file after the use_* function.}

\item{additional_args}{A named list of additional command line arguments to be
passed to the workflow. The names of the list represent the platforms (windows,
macos, or ubuntu), and the values are character vectors of arguments.
These additional arguments are executed after the step that checks out the
repository and before the step that sets up R.}
}
\description{
For more information, see \href{https://nmfs-fish-tools.github.io/ghactions4r/articles/set_up.html}{the setup details vignette} before
running this function. This workflow will assume the website is built from
a branch called \code{gh-pages}.
}
\examples{
\dontrun{
use_update_pkgdown(
  additional_args = list(
    ubuntu = c(
      "sudo apt-get update",
      "sudo apt-get install -y libcurl4-openssl-dev",
      "sudo add-apt-repository ppa:ubuntu-toolchain-r/test",
      "sudo apt-get install --only-upgrade libstdc++6"
    ),
    macos = c("brew install curl")
  )
)
}
}
