% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r_workflows.R
\name{use_update_roxygen_docs}
\alias{use_update_roxygen_docs}
\title{use workflow in current pkg to run devtools::document() and submit results as a PR}
\usage{
use_update_roxygen_docs(workflow_name = "call-update-docs.yml")
}
\arguments{
\item{workflow_name}{What to name the github action workflow locally. Must
have the file extension \code{.yml}}
}
\description{
use workflow in current pkg to run devtools::document() and submit results as a PR
}
